requirements.txt
numpy<1.25
pandas==2.1.0
Pillow==10.0.0
numba==0.57.1
librosa==0.10.1